refactor: move all browser related code from Vitest to Browser
