+++ packages/browser/package.json
@@ -75,7 +75,8 @@
     "@vitest/utils": "workspace:*",
     "magic-string": "^0.30.10",
     "msw": "^2.3.1",
-    "sirv": "^2.0.4"
+    "sirv": "^2.0.4",
+    "ws": "^8.17.1"
   },
   "devDependencies": {
     "@types/ws": "^8.5.10",
