+++ packages/browser/src/node/serverTester.ts
@@ -1,10 +1,10 @@
 import type { IncomingMessage, ServerResponse } from 'node:http'
 import { stringify } from 'flatted'
 import { replacer } from './utils'
-import type { BrowserServerState } from './state'
+import type { BrowserServer } from './server'
 
 export async function resolveTester(
-  state: BrowserServerState,
+  server: BrowserServer,
   url: URL,
   res: ServerResponse<IncomingMessage>,
 ): Promise<string> {
@@ -18,8 +18,9 @@ export async function resolveTester(
     )
   }
 
-  const { contextId, testFile } = state.resolveTesterUrl(url.pathname)
-  const project = state.project
+  const { contextId, testFile } = server.resolveTesterUrl(url.pathname)
+  const project = server.project
+  const state = server.state
   const testFiles = await project.globTestFiles()
   // if decoded test file is "__vitest_all__" or not in the list of known files, run all tests
   const tests
@@ -28,46 +29,46 @@ export async function resolveTester(
       ? '__vitest_browser_runner__.files'
       : JSON.stringify([testFile])
   const iframeId = JSON.stringify(testFile)
-  const files = project.browserState.get(contextId)?.files ?? []
+  const files = state.getContext(contextId)?.files ?? []
 
-  const injectorJs = typeof state.injectorJs === 'string'
-    ? state.injectorJs
-    : await state.injectorJs
+  const injectorJs = typeof server.injectorJs === 'string'
+    ? server.injectorJs
+    : await server.injectorJs
 
-  const config = state.getSerializableConfig()
+  const config = server.getSerializableConfig()
 
   const injector = replacer(injectorJs, {
-    __VITEST_PROVIDER__: JSON.stringify(project.browserProvider!.name),
+    __VITEST_PROVIDER__: JSON.stringify(server.provider.name),
     __VITEST_CONFIG__: JSON.stringify(config),
     __VITEST_FILES__: JSON.stringify(files),
     __VITEST_VITE_CONFIG__: JSON.stringify({
-      root: project.browser!.config.root,
+      root: server.vite.config.root,
     }),
     __VITEST_TYPE__: '"tester"',
     __VITEST_CONTEXT_ID__: JSON.stringify(contextId),
     __VITEST_PROVIDED_CONTEXT__: JSON.stringify(stringify(project.getProvidedContext())),
   })
 
-  if (!state.testerScripts) {
-    const testerScripts = await state.formatScripts(
+  if (!server.testerScripts) {
+    const testerScripts = await server.formatScripts(
       project.config.browser.testerScripts,
     )
     const clientScript = `<script type="module" src="${config.base || '/'}@vite/client"></script>`
-    const stateJs = typeof state.stateJs === 'string'
-      ? state.stateJs
-      : await state.stateJs
+    const stateJs = typeof server.stateJs === 'string'
+      ? server.stateJs
+      : await server.stateJs
     const stateScript = `<script type="module">${stateJs}</script>`
-    state.testerScripts = `${stateScript}${clientScript}${testerScripts}`
+    server.testerScripts = `${stateScript}${clientScript}${testerScripts}`
   }
 
-  const testerHtml = typeof state.testerHtml === 'string'
-    ? state.testerHtml
-    : await state.testerHtml
+  const testerHtml = typeof server.testerHtml === 'string'
+    ? server.testerHtml
+    : await server.testerHtml
 
   return replacer(testerHtml, {
-    __VITEST_FAVICON__: state.faviconUrl,
+    __VITEST_FAVICON__: server.faviconUrl,
     __VITEST_TITLE__: 'Vitest Browser Tester',
-    __VITEST_SCRIPTS__: state.testerScripts,
+    __VITEST_SCRIPTS__: server.testerScripts,
     __VITEST_INJECTOR__: injector,
     __VITEST_APPEND__:
       `<script type="module">
