+++ packages/vitest/src/node/automock.ts
