+++ packages/vitest/src/node/logger.ts
@@ -205,7 +205,7 @@ export class Logger {
       const name = project.getName()
       const output = project.isCore() ? '' : ` [${name}]`
 
-      const resolvedUrls = project.browser.resolvedUrls
+      const resolvedUrls = project.browser.vite.resolvedUrls
       const origin = resolvedUrls?.local[0] ?? resolvedUrls?.network[0]
       this.log(
         c.dim(
